Move EAC to Juvix.Core.EAC
