+++ src/Juvix/Core/EAC.hs
@@ -1,0 +1,14 @@
+module Juvix.Core.EAC
+  ( module Juvix.Core.EAC.EAC,
+    module Juvix.Core.EAC.Types,
+    module Juvix.Core.EAC.Parser,
+    module Juvix.Core.EAC.Check,
+    module Juvix.Core.EAC.Solve,
+  )
+where
+
+import Juvix.Core.EAC.Check
+import Juvix.Core.EAC.EAC
+import Juvix.Core.EAC.Parser
+import Juvix.Core.EAC.Solve
+import Juvix.Core.EAC.Types
