+++ src/Juvix/Core/EAC/Solve.hs
@@ -1,9 +1,9 @@
-module Juvix.EAC.Solve where
+module Juvix.Core.EAC.Solve where
 
 import qualified Data.Map as Map
 import qualified Data.Set as Set
 import GHC.Base (String)
-import Juvix.EAC.Types
+import Juvix.Core.EAC.Types
 import Juvix.Library hiding (link, reduce)
 import qualified Z3.Monad as Z3
