--- src/Juvix/EAC.hs
@@ -1,14 +1,0 @@
-module Juvix.EAC
-  ( module Juvix.EAC.EAC,
-    module Juvix.EAC.Types,
-    module Juvix.EAC.Parser,
-    module Juvix.EAC.Check,
-    module Juvix.EAC.Solve,
-  )
-where
-
-import Juvix.EAC.Check
-import Juvix.EAC.EAC
-import Juvix.EAC.Parser
-import Juvix.EAC.Solve
-import Juvix.EAC.Types
