+++ test/EAC.hs
@@ -1,6 +1,6 @@
 module EAC where
 
-import Juvix.EAC
+import Juvix.Core.EAC
 import Juvix.Library
 import Text.Parsec
