+++ test/Erasure.hs
@@ -5,8 +5,8 @@ import qualified Juvix.Backends.Graph as Graph
 import qualified Juvix.Backends.Maps as Maps ()
 import qualified Juvix.Bohm as Bohm
 import qualified Juvix.Core as Core
+import qualified Juvix.Core.EAC as EAC
 import qualified Juvix.Core.Erasure as Erasure
-import qualified Juvix.EAC as EAC
 import Juvix.Library hiding (identity)
 import qualified Juvix.Nets.Bohm as Bohm
 import qualified Test.Tasty as T
